changement des fonction du main dans un fichier qui lui est propre
