--- sys/ROS_Layer/dummy_file.md
@@ -1,4 +1,0 @@
-# Dummy
-
-
-
